# backend/handlers/pdf.go
package handlers

import (
	"bytes"
	_ "embed"
	"fmt"
	"os"
	"path/filepath"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/signintech/gopdf"
	"line-estimate-backend/models"
	"line-estimate-backend/services"
	"line-estimate-backend/utils"
)

//go:embed NotoSansJP-Regular.ttf
var notoSansJPFont []byte

// loadJapaneseFont loads Japanese font for PDF
func loadJapaneseFont(pdf *gopdf.GoPdf) error {
	// Use embedded font
	fontErr := pdf.AddTTFFontData("noto-sans", notoSansJPFont)
	if fontErr != nil {
		return fmt.Errorf("embedded font loading failed: %v", fontErr)
	}
	return nil
}

// GenerateEstimatePDF generates an estimate PDF from the provided data
// This is an internal function, not exposed as an API endpoint
func GenerateEstimatePDF(estimate *models.PDFEstimate) (*gopdf.GoPdf, error) {
	// Create a new PDF document
	pdf := &gopdf.GoPdf{}
	pdf.Start(gopdf.Config{PageSize: *gopdf.PageSizeA4})

	// Load Japanese font
	if err := loadJapaneseFont(pdf); err != nil {
		return nil, err
	}

	// First page
	pdf.AddPage()

	// Create helper
	helper := utils.NewPDFHelper(pdf)

	// Draw header
	if err := helper.DrawHeader(estimate); err != nil {
		return nil, err
	}

	// Draw customer info
	if err := helper.DrawCustomerInfo(estimate); err != nil {
		return nil, err
	}

	// Draw title
	if err := helper.DrawTitle(estimate.Title); err != nil {
		return nil, err
	}

	// Draw items table
	tableEndY, err := helper.DrawTable(estimate.Items, 320)
	if err != nil {
		return nil, err
	}

	// Draw totals
	if err := helper.DrawTotals(estimate, tableEndY+20); err != nil {
		return nil, err
	}

	// Draw remarks
	if len(estimate.Remarks) > 0 {
		if err := helper.DrawRemarks(estimate.Remarks, tableEndY+120); err != nil {
			return nil, err
		}
	}

	// Second page (for company seal)
	pdf.AddPage()
	pdf.SetX(450)
	pdf.SetY(100)
	if err := pdf.SetFont("noto-sans", "", 10); err != nil {
		return nil, err
	}
	pdf.Cell(nil, "社印")

	// Draw box for seal
	pdf.SetLineWidth(0.5)
	pdf.Rectangle(430, 120, 100, 100, "D", 0, 0)

	return pdf, nil
}

// CreateEstimatePDF godoc
// @Summary 見積もりPDFを生成
// @Description 見積もり情報からPDFを生成します
// @Tags Estimates
// @Accept json
// @Produce application/pdf
// @Param estimate body models.PDFEstimate true "見積もり情報"
// @Success 200 {file} binary
// @Failure 400 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Router /api/v1/estimates/pdf [post]
func CreateEstimatePDF(c *gin.Context) {
	var estimate models.PDFEstimate
	if err := c.ShouldBindJSON(&estimate); err != nil {
		utils.SendErrorResponse(c, 400, "無効なリクエストデータ: "+err.Error())
		return
	}

	// Generate PDF
	pdf, err := GenerateEstimatePDF(&estimate)
	if err != nil {
		utils.SendErrorResponse(c, 500, "PDF生成に失敗しました: "+err.Error())
		return
	}

	// PDFをバイト配列に変換
	var buf bytes.Buffer
	if err := pdf.Write(&buf); err != nil {
		utils.SendErrorResponse(c, 500, "PDFの書き込みに失敗しました: "+err.Error())
		return
	}

	// Generate unique filename with timestamp
	timestamp := time.Now().Format("20060102_150405")
	filename := fmt.Sprintf("estimate_%s_%s.pdf", estimate.EstimateNo, timestamp)

	// Check if local save only mode
	saveLocal := os.Getenv("SAVE_LOCAL_PDF")

	if saveLocal == "true" {
		// ローカル保存のみ
		pdfDir := "./pdfs"
		if err := os.MkdirAll(pdfDir, 0755); err != nil {
			utils.SendErrorResponse(c, 500, "PDFディレクトリの作成に失敗しました: "+err.Error())
			return
		}
		localPath := filepath.Join(pdfDir, filename)
		if err := pdf.WritePdf(localPath); err != nil {
			utils.SendErrorResponse(c, 500, "PDFのローカル保存に失敗しました: "+err.Error())
			return
		}

		// Return success response for local save
		utils.SuccessResponse(c, gin.H{
			"message":    "見積書PDFがローカルに正常に作成されました",
			"filename":   filename,
			"local_path": localPath,
		})
		return
	}

	// Google Driveサービスを初期化
	driveService, err := services.NewDriveService()
	if err != nil {
		utils.SendErrorResponse(c, 500, "Google Driveサービスの初期化に失敗しました: "+err.Error())
		return
	}

	// Google Driveにアップロード
	uploadedFile, err := driveService.UploadFile(filename, "application/pdf", buf.Bytes())
	if err != nil {
		utils.SendErrorResponse(c, 500, "Google Driveへのアップロードに失敗しました: "+err.Error())
		return
	}

	// Return success response
	utils.SuccessResponse(c, gin.H{
		"message":       "見積書PDFが正常に作成されました",
		"filename":      filename,
		"drive_file_id": uploadedFile.Id,
		"drive_url":     fmt.Sprintf("https://drive.google.com/file/d/%s/view", uploadedFile.Id),
	})
}

// CreatePDF godoc
// @Summary テストPDFを生成
// @Description 開発用のテストPDFを生成します
// @Tags Development
// @Accept json
// @Produce application/pdf
// @Success 200 {file} binary
// @Failure 500 {object} utils.ErrorResponse
// @Router /dev/create-pdf [get]
func CreatePDF(c *gin.Context) {
	// Create test estimate data
	testEstimate := &models.PDFEstimate{
		EstimateNo: "EST-20250425-001",
		IssueDate:  time.Now(),
		Customer: models.PDFCustomerInfo{
			CompanyName: "株式会社丸井",
			PostalCode:  "123-4567",
			Address:     "東京都新宿区○○1-2-3",
			Tel:         "03-1234-5678",
			Fax:         "03-8765-4321",
		},
		Recipient: "佐藤 様",
		Title:     "廃棄物処理に関する見積書",
		Items: []models.PDFLineItem{
			{
				Description: "オフィスチェア",
				Quantity:    10,
				UnitPrice:   1500,
				Amount:      15000,
			},
			{
				Description: "会議用テーブル",
				Quantity:    3,
				UnitPrice:   4000,
				Amount:      12000,
			},
			{
				Description: "収集運搬費",
				Quantity:    1,
				UnitPrice:   20000,
				Amount:      20000,
			},
		},
		SubTotal: 47000,
		TaxRate:  0.10,
		Tax:      4700,
		Total:    51700,
		Remarks: []string{
			"※お見積もりの有効期限は発行日より1ヶ月となります。",
			"※実際の廃棄物量により金額が変更となる場合がございます。",
			"※お支払い条件：作業完了後、請求書発行日より30日以内",
		},
	}

	// Generate PDF
	pdf, err := GenerateEstimatePDF(testEstimate)
	if err != nil {
		utils.SendErrorResponse(c, 500, "PDF生成に失敗しました: "+err.Error())
		return
	}

	// PDFをバイト配列に変換
	var buf bytes.Buffer
	if err := pdf.Write(&buf); err != nil {
		utils.SendErrorResponse(c, 500, "PDFの書き込みに失敗しました: "+err.Error())
		return
	}

	// Generate unique filename with timestamp
	timestamp := time.Now().Format("20060102_150405")
	filename := fmt.Sprintf("test_estimate_%s.pdf", timestamp)

	// Check if local save only mode
	saveLocal := os.Getenv("SAVE_LOCAL_PDF")

	if saveLocal == "true" {
		// ローカル保存のみ
		pdfDir := "./pdfs"
		if err := os.MkdirAll(pdfDir, 0755); err != nil {
			utils.SendErrorResponse(c, 500, "PDFディレクトリの作成に失敗しました: "+err.Error())
			return
		}
		localPath := filepath.Join(pdfDir, filename)
		if err := pdf.WritePdf(localPath); err != nil {
			utils.SendErrorResponse(c, 500, "PDFのローカル保存に失敗しました: "+err.Error())
			return
		}

		// Return success response for local save
		utils.SuccessResponse(c, gin.H{
			"message":    "テスト見積書PDFがローカルに正常に作成されました",
			"filename":   filename,
			"local_path": localPath,
		})
		return
	}

	// Google Driveサービスを初期化
	driveService, err := services.NewDriveService()
	if err != nil {
		utils.SendErrorResponse(c, 500, "Google Driveサービスの初期化に失敗しました: "+err.Error())
		return
	}

	// Google Driveにアップロード
	uploadedFile, err := driveService.UploadFile(filename, "application/pdf", buf.Bytes())
	if err != nil {
		utils.SendErrorResponse(c, 500, "Google Driveへのアップロードに失敗しました: "+err.Error())
		return
	}

	// Return success response
	utils.SuccessResponse(c, gin.H{
		"message":       "テスト見積書PDFが正常に作成されました",
		"filename":      filename,
		"drive_file_id": uploadedFile.Id,
		"drive_url":     fmt.Sprintf("https://drive.google.com/file/d/%s/view", uploadedFile.Id),
	})
}